Move Platform class into platform module due to partial import errors.
