+++ helpers/common.py
@@ -2,36 +2,14 @@
 
 import logging
 import os
-import platform
 import shutil
 import time
 from typing import Union
 
 from config import EMULATOR_CORE_AVG_FPS
+from helpers.platform import Platform
 from helpers.log import mod_fname
 logger = logging.getLogger(mod_fname(__file__))
-
-
-class Platform:
-    """Clean way to identify platform OS."""
-    MAC: str = "Darwin"
-    WINDOWS: str = "Windows"
-    LINUX: str = "Linux"
-
-    @staticmethod
-    def is_mac() -> bool:
-        """True if running on Macintosh; otherwise False."""
-        return platform.system() == Platform.MAC
-    
-    @staticmethod
-    def is_windows() -> bool:
-        """True if running on Windows; otherwise False."""
-        return platform.system() == Platform.WINDOWS
-    
-    @staticmethod
-    def is_linux() -> bool:
-        """True if running on Linux; otherwise False."""
-        return platform.system() == Platform.LINUX
 
 
 def delay(sec: float):
